several changes: moved for loop of event notification to be inside of run method
Commented out synchronized statements, because only single threaded approach is supported any more. When executed in single threaded mode, synchronization is not needed and is also problematic if the model execution thread hangs and keeps the lock. In such a case the fallback thread could not enter the code block.