bug fix: The stop method of the Timer plugin sent measured time to the output port in MEASURE_TIME mode. --> This can cause an endless loop under some circumstances.
-->
only send the measured time in case of a received stop event:
TimerInstance.receiveEvent calls TimeGenerator.stopAndSendData

TimerInstance.stop method calls TimeGenerator.stop method which does not send the data

*) Additionally removed synchronized keywords because the receiveEvent methods can only be called by the model executor thread and the TimeGenerator can only be called from TimerInstance
*) Cleaned up TimeGenerator.start method and remove endless loop waiting for runThread != null

Notes:
The test model still produces a ConcurrentModificationException when iterating through the event listeners in DefaultRuntimeEventTriggeredPort. But without a remarkable side effect. Nevertheless this should be investigated and exceptions should be catched within the for loop of event notification to not effect other notifications.