Fixes #101: Improved error messages!! DeploymentManager: now shows error dialog in plugin instantiation fails, and generally logs stack trace in case of an error. AREServices: made internal methods public so that someone could use them and get the exception back. Don't show reason to user but log stacktrace. AsapiSupport: deployModel now uses AREServices.deployModel
Big improvement in parsing old model file:
There was a RuntimeException when parsing old models when setting runtime property values. This was because in case of an old plugin, it's possible that a property value did not exist and hence is null. --> Caught throwing NullpointerException and added logging of ignored properties.