Fixes #152
When starting the ARE the webservices are turned on automatically.
If the flag --webservice is used it is simply ignored.

NOTE:
The file profile/services_websocketdemo.ini was renamed to profile/services_webservice.ini to better reflect the meaning of it.
This could cause an error message in APE-based projects which rely on that file.
How to solve: Simply remove the file name services_websocketdemo.ini when setting the property eu.asterics.ARE.ServicesFiles as the

How to turn off the webservice:
If you don't want the webservices to be started, edit the file profile/services_webservice.ini and comment out the respective asterics services.