Added ant target
installWebACS: to download the WebACS from github releases (https://github.com/asterics/WebACS/releases) and extract it to bin/ARE/web/WebACS/. The download is only performed if the file locally does not exist and the remote file is not newer. A tmp/ directory was introduced in the AsTERICS root directory to cache download files.
The clean target was updated to clean the tmp/ directory and the bin/ARE/web/WebACS directory